Patch libHSL upstream for ma57_get_factors
